(1) Modified run_network.cpp to read in 'n_steps' as a float and then cast to an int. This allows n_steps to be passed to run_network in exponential format.
(2) Fixed a bug in the PLA where exact-stochastic reaction times were
being calculated as -INFINITY if the rate = -0.0 (negative zero).

+++ bng2/Network3/src/run_network.cpp
@@ -308,7 +308,7 @@ int main(int argc, char *argv[]){
 	if ((argleft = argc - iarg) == 2) {
 		/* input is sample_time n_sample */
 		sample_time = atof(argv[iarg++]);
-		n_sample = atoi(argv[iarg++]);
+		n_sample = (int)atof(argv[iarg++]); // Read as float and cast to int to allow for exponential format
 	}
 	else {
 		/* input is t1 t2 ... tn */
